requirements.txt
python-dotenv==0.1.0
requests==2.21.0